+++ lib/src/ast.dart
@@ -8,3 +8,4 @@ export 'ast/functions.dart';
 export 'ast/calculus.dart';
 export 'ast/logic.dart';
 export 'ast/matrix.dart';
+export 'evaluation_result.dart';